Move HTTPClientMock into separate file
