+++ Sources/TestHelpers/HTTPClientMock.swift
@@ -1,0 +1,19 @@
+import Foundation
+import Gravatar
+
+struct HTTPClientMock: HTTPClient {
+    private let session: URLSessionMock
+
+    init(session: URLSessionMock) {
+        self.session = session
+    }
+
+    func fetchData(with request: URLRequest) async throws -> (Data, HTTPURLResponse) {
+        await session.update(request: request)
+        return (session.returnData, session.response)
+    }
+
+    func uploadData(with request: URLRequest, data: Data) async throws -> HTTPURLResponse {
+        session.response
+    }
+}
