+++ Sources/TestHelpers/HTTPURLResponse+Extension.swift
@@ -10,22 +10,3 @@ extension HTTPURLResponse {
         HTTPURLResponse(url: url!, statusCode: code, httpVersion: nil, headerFields: nil)!
     }
 }
-
-// MARK: - HTTPClient
-
-struct HTTPClientMock: HTTPClient {
-    private let session: URLSessionMock
-
-    init(session: URLSessionMock) {
-        self.session = session
-    }
-
-    func fetchData(with request: URLRequest) async throws -> (Data, HTTPURLResponse) {
-        await session.update(request: request)
-        return (session.returnData, session.response)
-    }
-
-    func uploadData(with request: URLRequest, data: Data) async throws -> HTTPURLResponse {
-        session.response
-    }
-}
